requirements.txt
langchain-core==1.2.14
langchain-openai==1.1.10
python-dotenv==1.2.1
langchain-community==0.4.1
langchain-classic==1.0.1
streamlit==1.54.0
yfinance==1.2.0
fastapi==0.115.8
uvicorn==0.34.0
requests==2.33.1
